gymAPI/requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.6.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.1
django-environ==0.9.0
django-extensions==3.2.0
django-money==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
environ==1.0
executing==0.9.1
Faker==13.15.1
future==0.18.2
gnureadline==8.1.2
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
ipython==8.4.0
ishell==0.1.8
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
protobuf==4.21.5
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-moneyed==2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
traitlets==5.3.0
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
